requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
chromadb==0.5.23
sentence-transformers==3.1.1
pydantic==2.9.2
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
requests==2.32.3
pdfminer.six==20231228
python-docx==1.1.2
readability-lxml==0.8.1
lxml==5.3.0
lxml_html_clean==0.3.1
